Reworking how launcher script works.
First of all, it is now generated by configure (autoconf) so the paths is
correct even if out-of-tree builds are used.

Secondly tool- and daemon arguments are separated by a -- argument (if not
provided all arguments are considered daemon arguments), so it is now possible
to pass arguments to valgrind.


svn path=/trunk/; revision=953